+++ src/main/java/de/adorsys/keycloak/config/provider/KeycloakProvider.java
@@ -35,6 +35,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.text.MessageFormat;
 import java.time.Duration;
@@ -89,6 +91,15 @@ public class KeycloakProvider implements AutoCloseable {
 
     public void refreshToken() {
         getInstance().tokenManager().refreshToken();
+    }
+
+    public <T> T getCustomApiProxy(Class<T> proxyClass) {
+        try {
+            URI uri = properties.getUrl().toURI();
+            return getInstance().proxy(proxyClass, uri);
+        } catch (URISyntaxException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     private Keycloak createKeycloak() {