+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for managing fine grained authorization rules with placeholders to reference identity providers by alias, realm role by name and groups by full path
 
 ## [4.8.1] - 2022-03-09
 
@@ -17,7 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Support for managing `Client Authorization Resources` like other resources by configuring `import.managed.client-authorization-resources=<full|no-delete>`. This prevents deletion of remote managed resources.
-- Support for managing fine granted authorization rules.
+- Support for managing fine granted authorization rules with placeholders to reference clients by client id.
 
 ### Changes